fix: modeles compatibles TiDB

+++ backend/models/Client.js
@@ -4,23 +4,23 @@ const User = require('./User');
 
 const Client = sequelize.define('Client', {
   id: {
-    type: DataTypes.INTEGER.UNSIGNED,
+    type: DataTypes.BIGINT.UNSIGNED,
     primaryKey: true,
     autoIncrement: true
   },
   userId: {
-    type: DataTypes.INTEGER.UNSIGNED,
+    type: DataTypes.BIGINT.UNSIGNED,
     allowNull: false,
     field: 'user_id'
   },
   telephone: { type: DataTypes.STRING, allowNull: true },
-  adresse: { type: DataTypes.STRING, allowNull: true }
+  adresse:   { type: DataTypes.STRING, allowNull: true }
 }, {
-  tableName: 'clients',
+  tableName:  'clients',
   timestamps: true
 });
 
 Client.belongsTo(User, { foreignKey: 'userId', as: 'user' });
-User.hasOne(Client, { foreignKey: 'userId', as: 'client' });
+User.hasOne(Client,    { foreignKey: 'userId', as: 'client' });
 
 module.exports = Client;
